deprcate run_cmd and run_cmd_qa & co, move them to easybuild._deprecated module
